Authorization  added for all interfaces which has write operations

+++ uniform-cms-openapi/src/main/java/com/nokia/ucms/openapi/v1/controller/ProjectCategoryApiController.java
@@ -1,12 +1,16 @@
 package com.nokia.ucms.openapi.v1.controller;
 
 import com.nokia.ucms.biz.entity.ProjectCategory;
+import com.nokia.ucms.biz.entity.ProjectColumn;
+import com.nokia.ucms.biz.entity.ProjectInfo;
+import com.nokia.ucms.biz.entity.User;
 import com.nokia.ucms.biz.service.ProjectCategoryService;
 import com.nokia.ucms.biz.service.ProjectInfoService;
 import com.nokia.ucms.common.controller.BaseController;
 import com.nokia.ucms.common.entity.ApiQueryResult;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,7 +25,7 @@ public class ProjectCategoryApiController extends BaseController
     private static Logger LOGGER = Logger.getLogger(ProjectCategoryApiController.class);
 
     @Autowired
-    private ProjectInfoService projectService;
+    private ProjectInfoService projectInfoService;
 
     @Autowired
     private ProjectCategoryService projectCategoryService;
@@ -49,24 +53,34 @@ public class ProjectCategoryApiController extends BaseController
     }
 
     @RequestMapping(path="", method= RequestMethod.POST)
+    @PreAuthorize("isAuthenticated()")
     public @ResponseBody ApiQueryResult<ProjectCategory> createProjectCategory(
             @PathVariable Integer projectId, @RequestBody ProjectCategory projectCategory)
     {
         if (LOGGER.isDebugEnabled())
             LOGGER.debug(String.format("Enter createProjectCategory - [projectId: %d]", projectId));
 
-        ProjectCategory category = this.projectCategoryService.addProjectCategory(projectId, projectCategory);
-        if (category != null)
+        ProjectInfo entity = this.projectInfoService.getProjectById(projectId);
+        if (entity.getOwner() != null && entity.getOwner().getId().equals(((User)getAuthenticatedPrinciple()).getId()))
         {
-            return new ApiQueryResult<ProjectCategory>(category);
+            ProjectCategory category = this.projectCategoryService.addProjectCategory(projectId, projectCategory);
+            if (category != null)
+            {
+                return new ApiQueryResult<ProjectCategory>(category);
+            }
+            else
+            {
+                return new ApiQueryResult<ProjectCategory>(false, null, "Failed to create project category");
+            }
         }
         else
         {
-            return new ApiQueryResult<ProjectCategory>(false, null, "Failed to create project category");
+            return new ApiQueryResult<ProjectCategory>(false, null, "No permission to create category in this project");
         }
     }
 
     @RequestMapping(path="/{categoryId}", method= RequestMethod.PATCH)
+    @PreAuthorize("isAuthenticated()")
     public @ResponseBody ApiQueryResult<ProjectCategory> updateProjectCategory(
             @PathVariable Integer projectId,
             @PathVariable Integer categoryId,
@@ -77,17 +91,25 @@ public class ProjectCategoryApiController extends BaseController
 
         if (projectCategory != null)
         {
-            // overwrite with given project id anyway to avoid incorrect update
-            projectCategory.setProjectId(projectId);
-
-            ProjectCategory category = this.projectCategoryService.updateProjectCategory(projectId, categoryId, projectCategory);
-            if (category != null)
+            ProjectInfo entity = this.projectInfoService.getProjectById(projectId);
+            if (entity.getOwner() != null && entity.getOwner().getId().equals(((User)getAuthenticatedPrinciple()).getId()))
             {
-                return new ApiQueryResult<ProjectCategory>(category);
+                // overwrite with given project id anyway to avoid incorrect update
+                projectCategory.setProjectId(projectId);
+
+                ProjectCategory category = this.projectCategoryService.updateProjectCategory(projectId, categoryId, projectCategory);
+                if (category != null)
+                {
+                    return new ApiQueryResult<ProjectCategory>(category);
+                }
+                else
+                {
+                    return new ApiQueryResult<ProjectCategory>(false, null, "Failed to update project category: " + projectCategory);
+                }
             }
             else
             {
-                return new ApiQueryResult<ProjectCategory>(false, null, "Failed to update project category: " + projectCategory);
+                return new ApiQueryResult<ProjectCategory>(false, null, "No permission to update this category");
             }
         }
 
@@ -95,14 +117,23 @@ public class ProjectCategoryApiController extends BaseController
     }
 
     @RequestMapping(path="/{categoryId}", method= RequestMethod.DELETE)
+    @PreAuthorize("isAuthenticated()")
     public @ResponseBody ApiQueryResult<Object> deleteProjectCategory(
             @PathVariable Integer projectId, @PathVariable Integer categoryId)
     {
         if (LOGGER.isDebugEnabled())
             LOGGER.debug(String.format("Enter deleteProjectCategory - [projectId: %d, categoryId: %d]", projectId, categoryId));
 
-        boolean result = this.projectCategoryService.removeProjectCategory(categoryId);
-        return new ApiQueryResult<Object>(result, null);
+        ProjectInfo entity = this.projectInfoService.getProjectById(projectId);
+        if (entity.getOwner() != null && entity.getOwner().getId().equals(((User)getAuthenticatedPrinciple()).getId()))
+        {
+            boolean result = this.projectCategoryService.removeProjectCategory(categoryId);
+            return new ApiQueryResult<Object>(result, null);
+        }
+        else
+        {
+            return new ApiQueryResult<Object>(false, null, "No permission to delete this category");
+        }
     }
 
     protected String getModulePath ()
